[Oztechan/CCC#4198] Move to GitLive analytics and use an implementation in common source
