+++ client/core/analytics/client-core-analytics.gradle.kts
@@ -15,12 +15,10 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(libs.common.koinCore)
+            implementation(libs.common.gitliveAnalytics)
         }
         commonTest.dependencies {
             implementation(libs.common.test)
-        }
-        androidMain.dependencies {
-            implementation(libs.android.firebaseAnalytics)
         }
     }
 }
