--- client/core/analytics/src/iosMain/kotlin/com/oztechan/ccc/client/core/analytics/di/ClientCoreAnalyticsModule.ios.kt
@@ -1,8 +1,0 @@
-package com.oztechan.ccc.client.core.analytics.di
-
-import com.oztechan.ccc.client.core.analytics.AnalyticsManager
-import org.koin.dsl.module
-
-fun getClientCoreAnalyticsModule(analyticsManager: AnalyticsManager) = module {
-    single { analyticsManager }
-}
