+++ gradle/libs.versions.toml
@@ -17,7 +17,6 @@ koinAndroid = "4.0.2"
 koinKtor = "4.0.2"
 ktor = "2.3.13"
 multiplatformSettings = "1.3.0"
-firebaseAnalytics = "22.3.0"
 firebaseRemoteConfig = "22.1.0"
 googleServices = "4.4.2"
 firebasePer = "21.0.4"
@@ -43,6 +42,7 @@ rootBeer = "0.1.1"
 anrWatchDog = "1.4.0"
 kermit = "2.0.5"
 konsist = "0.17.3"
+gitlive = "2.1.0"
 
 [libraries]
 # COMMON
@@ -63,6 +63,7 @@ common-mokoResources = { module = "dev.icerock.moko:resources", version.ref = "m
 common-kermit = { module = "co.touchlab:kermit", version.ref = "kermit" }
 common-detektFormatting = { module = "io.gitlab.arturbosch.detekt:detekt-formatting", version.ref = "detekt" }
 common-konsist = { module = "com.lemonappdev:konsist", version.ref = "konsist" }
+common-gitliveAnalytics = { module = "dev.gitlive:firebase-analytics", version.ref = "gitlive" }
 
 # ANDROID
 android-androidDesugaring = { module = "com.android.tools:desugar_jdk_libs", version.ref = "androidDesugaring" }
@@ -74,7 +75,6 @@ android-navigation = { module = "androidx.navigation:navigation-fragment-ktx", v
 
 android-koinCompose = { module = "io.insert-koin:koin-androidx-compose", version.ref = "koinCompose" }
 android-koinAndroid = { module = "io.insert-koin:koin-android", version.ref = "koinAndroid" }
-android-firebaseAnalytics = { module = "com.google.firebase:firebase-analytics", version.ref = "firebaseAnalytics" }
 android-firebaseRemoteConfig = { module = "com.google.firebase:firebase-config", version.ref = "firebaseRemoteConfig" }
 android-firebasePer = { module = "com.google.firebase:firebase-perf", version.ref = "firebasePer" }
 android-lifecycleViewmodel = { module = "androidx.lifecycle:lifecycle-viewmodel-ktx", version.ref = "lifecycle" }
